+++ src/DetailsView/components/scan-incomplete-warning.tsx
